Set Step Reward
- 기존의 step reward 식 변경 완료

+++ Assets/Scripts/Agents/Data/ParkingAgentRewardData.cs
@@ -11,9 +11,9 @@ namespace AutonomousParking.Agents.Data
     {
         [field: Header("Target Reaching Rewards")]
         [field: SerializeField] // 최소 활동에 대한 보상 값입니다. x축으로 평행이동 하는 값
-        public float Shift { get; private set; }
+        public int Shift { get; private set; }
         [field: SerializeField] // 최대 한계점을 설정
-        public float StepRewardThreshold { get; private set; }
+        public int StepRewardThreshold { get; private set; }
 
         [field: SerializeField] // 목표까지의 거리를 줄였을 때 단계별로 주어지는 최대 보상입니다.
         public float MaxRewardForDecreasingDistanceToTargetPerStep { get; private set; }
